🧪  [contracts] dissolve test constants/utils into main ones
