--- modules/contracts/src.ts/tests/index.ts
@@ -1,2 +1,0 @@
-export * from "./constants";
-export * from "./utils";
